enhance(settings): add new fields to admin settings command (#602)

+++ command/deployment/view.go
@@ -40,7 +40,7 @@ var CommandView = &cli.Command{
 
 		// Deployment Flags
 
-		&cli.IntFlag{
+		&cli.Int64Flag{
 			Sources: cli.EnvVars("VELA_DEPLOYMENT", "DEPLOYMENT_NUMBER"),
 			Name:    "deployment",
 			Aliases: []string{"d", "number", "dn"},
@@ -103,7 +103,7 @@ func view(ctx context.Context, c *cli.Command) error {
 		Action: internal.ActionView,
 		Org:    c.String(internal.FlagOrg),
 		Repo:   c.String(internal.FlagRepo),
-		Number: c.Int("deployment"),
+		Number: c.Int64("deployment"),
 		Output: c.String(internal.FlagOutput),
 		Color:  output.ColorOptionsFromCLIContext(c),
 	}
